fix(starry-kernel): keep rknpu devfs in std builds
