+++ os/StarryOS/kernel/src/pseudofs/dev/dma_heap.rs
@@ -41,24 +41,15 @@ impl DeviceOps for DmaHeapSystem {
     fn ioctl(&self, cmd: u32, arg: usize) -> VfsResult<usize> {
         warn!("dma_heap: ioctl called cmd={:#x}, arg={:#x}", cmd, arg);
 
-        // Handle common DMA heap ioctls
-        match cmd {
-            // For now, we just return success for all ioctls and zero the first u32
-            // if arg is a user pointer, similar to rknpu implementation
-            _ => {
-                // Best-effort: if arg is a user pointer, zero the first u32 there so
-                // user-space doesn't read uninitialized memory
-                if arg != 0 {
-                    // write a safe default (0) to the user pointer
-                    // Use vm_write to safely write across the VM boundary.
-                    if let Err(e) = (arg as *mut u32).vm_write(0u32) {
-                        warn!("dma_heap: ioctl vm_write failed: {:?}", e);
-                        return Err(VfsError::InvalidInput);
-                    }
-                }
-                Ok(0)
-            }
+        // For now, return success for all ioctls and zero the first u32 if arg
+        // is a user pointer, similar to the rknpu implementation.
+        if arg != 0
+            && let Err(e) = (arg as *mut u32).vm_write(0u32)
+        {
+            warn!("dma_heap: ioctl vm_write failed: {:?}", e);
+            return Err(VfsError::InvalidInput);
         }
+        Ok(0)
     }
 
     fn as_any(&self) -> &dyn Any {
