+++ os/StarryOS/kernel/src/pseudofs/dev/mod.rs
@@ -1,9 +1,15 @@
 //! Special devices
 
 mod card0;
-#[cfg(all(feature = "rknpu", not(any(windows, unix))))]
+#[cfg(all(
+    feature = "rknpu",
+    any(not(any(windows, unix)), feature = "std-compat")
+))]
 mod card1;
-#[cfg(all(feature = "rknpu", not(any(windows, unix))))]
+#[cfg(all(
+    feature = "rknpu",
+    any(not(any(windows, unix)), feature = "std-compat")
+))]
 mod dma_heap;
 mod drm;
 #[cfg(feature = "input")]
@@ -22,9 +28,10 @@ pub use r#loop::LoopDevice;
 pub mod ion;
 #[cfg(feature = "memtrack")]
 mod memtrack;
-#[cfg(all(feature = "rknpu", not(any(windows, unix))))]
-mod rknpu_card;
-#[cfg(all(feature = "rknpu", not(any(windows, unix))))]
+#[cfg(all(
+    feature = "rknpu",
+    any(not(any(windows, unix)), feature = "std-compat")
+))]
 mod rknpu_drm;
 mod rtc;
 #[cfg(all(feature = "sg2002", not(feature = "plat-dyn")))]
@@ -367,7 +374,10 @@ fn builder(fs: Arc<SimpleFs>) -> DirMaker {
         ),
     );
 
-    #[cfg(all(feature = "rknpu", not(any(windows, unix))))]
+    #[cfg(all(
+        feature = "rknpu",
+        any(not(any(windows, unix)), feature = "std-compat")
+    ))]
     {
         // DMA heap devices (rknpu only)
         let mut dma_heap_dir = DirMapping::new();
