+++ os/StarryOS/starryos/Cargo.toml
@@ -45,6 +45,7 @@ k230 = [
 plat-dyn = ["dep:axplat-dyn", "starry-kernel/plat-dyn"]
 
 rknpu = ["ax-driver/rknpu", "starry-kernel/rknpu"]
+std-compat = ["starry-kernel/std-compat"]
 vf2 = ["ax-hal/riscv64-visionfive2"]
 
 [[bin]]
